+++ src/defaults.ts
@@ -4,12 +4,13 @@ import { loggerIsEnabled } from './logger';
 const options: Options = {
     source: null,
     calculateZMethod: 'getFeatureInfo',
+    tilesResolution: 'current',
+    bands: 4,
     samples: 50,
     smooth: 0,
     sampleSizeArea: 'auto',
     noDataValue: -10000,
-    verbose: loggerIsEnabled,
-    tilesResolution: 'current'
+    verbose: loggerIsEnabled
 };
 
 export default options;